Fixed a bug in an example + option for config edit
- Fixed #2

- Added some functions for configuration.
Old `getConfiguration()` is now `getRawConfiguration` (returns an `uint8_t`),
the new `getConfiguration()` returns a `MemoryConfig` structure.

`setConfiguration()` allows user to tune all parameters at once. It's also
the only method to enable the Configuration Register Lock (it makes this change a
two-stage process).
After using `setConfiguration()`, you need to send it to the 24CSM01 with
`updateConfiguration()`.